server/requirements.txt
asgiref==3.3.4
Django==3.2.4
djangorestframework==3.12.4
numpy==1.21.0
Pillow==8.2.0
Pygments==2.9.0
pytz==2021.1
sqlparse==0.4.1
